ok tested all quiz api endpoints and edge cases with some modifications

+++ src/controllers/Quiz.ts
@@ -5,22 +5,33 @@ import User from '../models/User';
 import { verifyToken } from '../library/Token';
 
 const createQuiz = async (req: Request, res: Response, next: NextFunction) => {
-    try {
-        const { name, description = null, datetime: dateString, user } = req.body;
-        const datetime = new Date(dateString);
+    const authHeader = req.headers.authorization;
+    if (authHeader) {
+        const token = authHeader.split(' ')[1];
+        try {
+            const decoded_user = verifyToken(token);
+            const { name, description = null, datetime: dateString } = req.body;
+            const datetime = new Date(dateString);
+            const user = await User.findOne({ username: decoded_user.username });
 
-        const quiz = new Quiz({
-            _id: new mongoose.Types.ObjectId(),
-            name,
-            description,
-            datetime,
-            user
-        });
-
-        const savedQuiz = await quiz.save();
-        return res.status(201).json({ quiz: savedQuiz });
-    } catch (error) {
-        return res.status(500).json({ message: 'Internal server error' });
+            if (user) {
+                const quiz = new Quiz({
+                    _id: new mongoose.Types.ObjectId(),
+                    name,
+                    description,
+                    datetime,
+                    user: user._id
+                });
+                const savedQuiz = await quiz.save();
+                return res.status(201).json({ quiz: savedQuiz });
+            } else {
+                return res.status(404).json({ message: 'User not found' });
+            }
+        } catch (error) {
+            return res.status(401).json({ message: 'Invalid token' });
+        }
+    } else {
+        return res.status(401).json({ message: 'Missing authorization header' });
     }
 };
 
@@ -30,7 +41,7 @@ const getQuiz = async (req: Request, res: Response, next: NextFunction) => {
         try {
             const quiz_id = req.params.id;
             const token = authHeader.split(' ')[1];
-            const user = verifyToken(token);
+            verifyToken(token);
 
             const quiz = await Quiz.findById(quiz_id);
             if (quiz) {
@@ -55,14 +66,14 @@ const updateQuiz = async (req: Request, res: Response, next: NextFunction) => {
             if (user) {
                 const quiz_id = req.params.id;
                 const old_quiz = await Quiz.findById(quiz_id);
-                const user_object = await User.findOne({ username: user.username })
+                const user_object = await User.findOne({ username: user.username });
+
                 if (old_quiz && user_object) {
-                    console.log(old_quiz.user, user_object._id)
                     if (!old_quiz.user.equals(user_object._id)) {
                         return res.status(401).json({ message: 'Unauthorized' });
                     }
 
-                    const { name = old_quiz.name, description = old_quiz, datetime = old_quiz.datetime } = req.body;
+                    const { name = old_quiz.name, description = old_quiz.description, datetime = old_quiz.datetime } = req.body;
                     const quiz = await Quiz.findByIdAndUpdate(quiz_id, { name, description, datetime }, { new: true });
                     res.status(200).json({ quiz });
                 } else {
@@ -84,10 +95,12 @@ const deleteQuiz = async (req: Request, res: Response, next: NextFunction) => {
             const token = authHeader.split(' ')[1];
             const user = verifyToken(token);
             if (user) {
-                const quiz_id = req.params.id;
+                const quiz_id = req.params.id.toString();
+
                 const old_quiz = await Quiz.findById(quiz_id);
-                if (old_quiz) {
-                    if (old_quiz.user.username !== user.username) {
+                const user_object = await User.findOne({ username: user.username });
+                if (old_quiz && user_object) {
+                    if (!old_quiz.user.equals(user_object._id)) {
                         return res.status(401).json({ message: 'Unauthorized' });
                     }
                     const quiz = await Quiz.findByIdAndDelete(quiz_id);
@@ -96,6 +109,8 @@ const deleteQuiz = async (req: Request, res: Response, next: NextFunction) => {
                     } else {
                         res.status(404).json({ message: 'Quiz not found' });
                     }
+                } else {
+                    res.status(404).json({ message: 'Quiz not found' });
                 }
             }
         } catch (error) {
